requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
langchain==0.2.13
langchain-community==0.2.12
langchain-openai==0.1.21
python-dotenv==1.0.1
streamlit==1.37.1
langchain-google-genai==1.0.8